fixed #32, modern architecture
