+++ src/test/kotlin/EntryMetaParseSpec.kt
@@ -2,6 +2,7 @@ package jp.yappo.pologen
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
+import jp.yappo.pologen.domain.config.EntryMeta
 import kotlin.io.path.createTempDirectory
 import kotlin.io.path.writeText
