+++ src/test/kotlin/HtmlUtilSpec.kt
@@ -2,6 +2,7 @@ package jp.yappo.pologen
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
+import jp.yappo.pologen.application.util.stripHtml
 
 class HtmlUtilSpec : FunSpec({
     test("stripHtml removes tags and trims") {
